readme added William Mills, refactored rules.py to generative_rules.py, pycache removed
