+++ generative_rules.py
